ddd模式-单机事务处理

+++ demo-ddd-interface/src/main/java/org/example/api/GroupApi.java
@@ -11,7 +11,7 @@ public interface GroupApi {
     Long updateGroup(GroupUpdateCmd updateCmd);
 
 
-    Integer deleteGroup(GroupDeleteCmd deleteCmd);
+    Long deleteGroup(GroupDeleteCmd deleteCmd);
 
 
 }
